docs: add heading link (#39388)
* docs: add heading link

* refactor: global styles
