+++ .dumi/theme/common/styles/BrowserMockup.tsx
@@ -1,0 +1,62 @@
+import { css, Global } from '@emotion/react';
+import React from 'react';
+
+export default () => (
+  <Global
+    styles={css`
+          /* Browser mockup code
+ * Contribute: https://gist.github.com/jarthod/8719db9fef8deb937f4f
+ * Live example: https://updown.io
+ */
+
+          .browser-mockup {
+            position: relative;
+            border-top: 2em solid rgba(230, 230, 230, 0.7);
+            border-radius: 3px 3px 0 0;
+            box-shadow: 0 0.1em 0.5em 0 rgba(0, 0, 0, 0.28);
+          }
+
+          .browser-mockup::before {
+            position: absolute;
+            top: -1.25em;
+            left: 1em;
+            display: block;
+            width: 0.5em;
+            height: 0.5em;
+            background-color: #f44;
+            border-radius: 50%;
+            box-shadow: 0 0 0 2px #f44, 1.5em 0 0 2px #9b3, 3em 0 0 2px #fb5;
+            content: '';
+          }
+
+          .browser-mockup.with-tab::after {
+            position: absolute;
+            top: -2em;
+            left: 5.5em;
+            display: block;
+            width: 20%;
+            height: 0;
+            border-right: 0.8em solid transparent;
+            border-bottom: 2em solid white;
+            border-left: 0.8em solid transparent;
+            content: '';
+          }
+
+          .browser-mockup.with-url::after {
+            position: absolute;
+            top: -1.6em;
+            left: 5.5em;
+            display: block;
+            width: ~'calc(100% - 6em)';
+            height: 1.2em;
+            background-color: white;
+            border-radius: 2px;
+            content: '';
+          }
+
+          .browser-mockup > * {
+            display: block;
+          }
+        `}
+  />
+);
