+++ .dumi/theme/common/styles/Reset.tsx
@@ -1,0 +1,63 @@
+import { css, Global } from '@emotion/react';
+import React from 'react';
+import useSiteToken from '../../../hooks/useSiteToken';
+
+export default () => {
+  const { token } = useSiteToken();
+
+  return (
+    <Global
+      styles={css`
+          @font-face {
+            font-weight: normal;
+            font-family: PuHuiTi;
+            src: url('//at.alicdn.com/t/webfont_6e11e43nfj.woff2') format('woff2'),
+              url('//at.alicdn.com/t/webfont_6e11e43nfj.woff') format('woff'),
+              /* chrome、firefox */ url('//at.alicdn.com/t/webfont_6e11e43nfj.ttf')
+                format('truetype'); /* chrome、firefox、opera、Safari, Android, iOS 4.2+ */
+            font-display: swap;
+          }
+
+          @font-face {
+            font-weight: bold;
+            font-family: PuHuiTi;
+            src: url('//at.alicdn.com/t/webfont_exesdog9toj.woff2') format('woff2'),
+              url('//at.alicdn.com/t/webfont_exesdog9toj.woff') format('woff'),
+              /* chrome、firefox */ url('//at.alicdn.com/t/webfont_exesdog9toj.ttf')
+                format('truetype'); /* chrome、firefox、opera、Safari, Android, iOS 4.2+ */
+            font-display: swap;
+          }
+
+          // 组件丰富，选用自如定制主题随心所欲设计语言与研发框架1234567890 QWERTYUIOPLKJHGFDSAZXCVBNM,.mnbvcxzasdfghjklpoiuytrewq
+          /* CDN 服务仅供平台体验和调试使用，平台不承诺服务的稳定性，企业客户需下载字体包自行发布使用并做好备份。 */
+          @font-face {
+            font-weight: 900;
+            font-family: 'AliPuHui';
+            src: url('//at.alicdn.com/wf/webfont/exMpJIukiCms/Gsw2PSKrftc1yNWMNlXgw.woff2')
+                format('woff2'),
+              url('//at.alicdn.com/wf/webfont/exMpJIukiCms/vtu73by4O2gEBcvBuLgeu.woff')
+                format('woff');
+            font-display: swap;
+          }
+
+          html {
+            direction: initial;
+
+            &.rtl {
+              direction: rtl;
+            }
+          }
+
+          body {
+            overflow-x: hidden;
+            color: ${token.colorText};
+            font-size: ${token.fontSize}px;
+            font-family: ${token.fontFamily};
+            line-height: ${token.lineHeight};
+            background: ${token.colorBgContainer};
+            transition: background 1s cubic-bezier(0.075, 0.82, 0.165, 1);
+          }
+        `}
+    />
+  );
+};
